DP-17062: Automate Github pr and release description (#900)
* setup test

* mock data

* wip

* wip

* wip

* fix

* add notes

* test it here!!!!

* or see the changelogs compilation here!!

* uncomment everything

* remove test data

* Update scripts/release-branch.js

* Update scripts/release-branch.js

* wip

* add description to pr

* fix

* add release tag node script

* wip

* pull out shared vars

* wip

* wip

* consolidate changelogs

* fix

* remove github tag shell script

* fix

* separate compile vs consolidate

* fix

* update circle

* generate temp logs

* read temp file

* read temp file

* remove templogs

* fix

* fix

* add notes

* remove test data

* fix circle

* test dependabot changelog and make issue optional

* format

* remove test

* add note

* fix circle steps

* update var name

* form new logs with title from changelog.md

* wip

* remove adding temp file

* clean up

* clean up

* update circle config

* clean up
