+++ .gitignore
@@ -5,3 +5,6 @@
 node_modules
 
 .env
+
+# Temporary changelog compilation
+tempLogs.json
